Don't insert into non-existent columns on import
Previously, when importing messages, sms-ie tried to insert all values
included in the supplied JSON (with certain enumerated exceptions) into
the local SMS and MMS databases, and would thus crash if some fields
(columns) present in the JSON did not exist in those databases (due to
changes in the database structure between the Android version of the
system from which the messages were exported and that of the one to
which they were being imported). Now, sms-ie drops every field not
present in the columns of the current databases.

 * Update copyright notice for 2022

 * Update Kotlin Gradle plugin

Closes: #11