Implement progress reporting
Implement progress reporting, by progress bars and status text, for
exporting and importing.

Note that during import, the app does not know in advance the total
number of messages or calls to be imported, since the JSON is streamed
sequentially, so the app can only report the number imported so far, but
not the percentage of the total.