Implement progress reporting
Implement progress reporting for exporting and importing.

Note that when importing, the app does not know in advance the total
number of messages or calls to be imported, since the JSON input is
processed sequentially (streamed), so the app can therefore only report
the number of messages or calls imported so far, but not the percentage
of the total number to be imported.

Closes: #24
https://github.com/tmo1/sms-ie/issues/24