Try to run scheduled exports in foreground service where possible
Android 14 introduced a change in the CachedAppOptimizer component that
kills apps that perform too many binder transactions in the background
[1]. This can happen where there are many messages to export (2000 or so
from empirical testing).

This restriction does not apply to foreground services, so this commit
adds support for opportunistically running ExportWorker inside
WorkManager's foreground service. Unfortunately, on Android 12+, this is
not allowed by default because ExportWorker's startup method of using a
delay on a OneTimeWorkRequestBuilder is not one of the exemptions for
running a foreground service.

To allow a foreground service to run, the user must disable Android's
battery optimizations for the app. This commit adds a toggle to do so in
the sms-ie's own settings so the user doesn't need to dig into Android's
Settings to find the option. If foreground service access is not
available (because eg. battery optimizations are enabled or the app is
running on an old version of Android), then the worker will fall back to
running in a plain old background service like before.

[1] https://android.googlesource.com/platform/frameworks/base.git/+/71d75c09b9a06732a6edb4d1488d2aa3eb779e14%5E%21/
[2] https://developer.android.com/guide/components/foreground-services#background-start-restriction-exemptions

Fixes: #129

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>