Report import and export progress through callbacks
This is the first step in decoupling the UI from the operations. An
upcoming commit will move the invocation of each operation to the worker
and report progress through the WorkInfo API.

Note that this commit temporarily breaks the functionality for hiding
the progress bar after an operation is complete.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>