ImportExportWorker: Add support for running the manual operations
By default, if no arguments are provided to the worker, it will perform
a scheduled export. Otherwise, the action argument determines which
(manual) action to perform. Representing the scheduled export by the
lack of an action argument is done to avoid needing to reschedule the
next scheduled run with a new WorkRequest when upgrading from a previous
version of sms-ie. By keeping the scheduled WorkRequest as simple as
possible, future refactoring is also simpler.

Aside from the action argument, there's also a file argument for the
manual import and export (but not wipe) actions. This is just the URI
that the user selected.

All of these operations use the previously introduced infrastructure for
reporting progress and success/failure entirely through the androidx
work APIs.

This commit also adds new strings so that we have text for the
in-progress and failure states of every operation.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>