Report all operation errors through exceptions
This way, errors of any kind will be funneled through the work Result
where it can be used for a notification or observed by MainActivity (in
an upcoming commit).

Some of these errors were intentionally meant to be user-facing. For
these errors, a new UserFriendlyException has been introduced. When this
is thrown, the error message will include both the friendly error
message and the first cause of the exception, if one exists. This
results in the same error messages as we do now.

With this change, none of the operations need the Context to be an
Activity context anymore, since they don't directly launch an
AlertDialog. The dialog handling will be readded in an upcoming commit
when MainActivity switches over to using the worker for manual
operations.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>