MainActivity: Update progress and status based on worker state
The activity will now observe the workers and update the progress bar
and status text based on their state. With this change, MainActivity is
now also able to show the state of the scheduled export while it is
running.

This commit doesn't make MainActivity launch the worker for manual
operations yet. That will be done in an upcoming commit.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>