MainActivity: Launch all manual actions through the worker
This is the final step is decoupling the activity from the operations
and using the worker. In a previous commit, MainActivity has already
been updated to observe the state of the workers, so MainActivity can
now detach and reattach to the workers without impacting the actual
operations. The user can even swipe away the activity.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>