Add support for cancelling workers
This uses the androidx work library's builtin mechanism for signaling
the cancellation of the worker. Each operation will periodically check
if it should stop running. This just requires an ensureActive() call if
the function does not frequently call any other suspend function. Since
cancellation is just throwing a CancellationException, the cleanup is
orderly (eg. all finally blocks will run).

The only operation that cannot be cancelled is the message wiping
operation. That uses ContentResolver.delete(), which has no variant that
accepts a CancellationSignal. The only way to interrupt that is to kill
the app, so cancellation is blocked for this operation instead.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>