ImportExportWorker: Save logcat of latest operation
This updates the worker to spawn and kill logcat and the start and end
of every operation. The output is written to sms-ie's external files
directory, which is guaranteed to exist and be accessible using regular
filesystem APIs. This is useful when an error is not easily reproducible
and the user isn't able to get to their computer and grab a logcat via
adb before the logs rotate.

When an error occurs, the error dialog has a new button to open the
directory containing the log file in the system file manager
(DocumentsUI). DocumentsUI and USB access are the only two ways to
access the external files directory in newer versions of Android.
However, DocumentsUI is full featured enough that the user can share the
log file or copy it elsewhere.

The downside of these logs compared to `adb logcat` is that they will
only contain our own logs. Logs from other apps or other system
components (eg. telephony provider) are excluded.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>