Only allow background services notifications to alert once
This fixes the device playing the notification source or vibrating every
time the notification is updated (for progress bar changes). I didn't
catch this when implementing #268 because I had a Pixel Watch connected,
which always forces the `setOnlyAlertOnce(true)` behavior.

Additionally, this changes the default importance of the background
services notifications to low (referred to as "Silent" in Android's UI).
Even with the original bug fixed, it doesn't make much sense to play a
sound or vibrate one time when an operation starts. This change in
defaults only takes effect for new installs. It is not possible to
change existing notification channel settings.

Fixes: #276

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>