ImportExportWorker: Handle setForeground() notification update failures
For whatever reason, the androidx work library starts a new background
service to display the notification each time it's updated. If the
worker was launched from MainActivity and the screen turns off, the
library's Context.startService() call will fail with the
BackgroundServiceStartNotAllowedException. androidx's foreground service
is still running when this happens, so perhaps this is an Android bug
where it loses track of the fact that the app is still in the
"foreground".

In any case, if we can no longer update the notification via
setForeground() calls, we can still update it via NotificationManager
directly.

This commit also adds handling for IllegalStateException since Android
versions prior to 12 throw that exception instead of
ForegroundServiceStartNotAllowedException.

Fixes: #277

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>